[Feature] Manual definition of form element order when using annotations
The order of form elements when generating a form using annotations is
determined by the order of the corresponding get- and is- functions
inside the model class source code.

Previously the order of form elements when using annotations inside a
class to generate a flux form was determined by the alphabetical order
of the corresponding get- and is- functions.